pyproject.toml
[project]
name = "admissionbot-rag"
requires-python = ">=3.9"

[tool.poetry]
name = "admissionbot-rag"
version = "0.1.0"
description = ""
authors = ["Peemmaphat Sripongsai <83326313+SupeemAFK@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-text-splitters = "^0.3.6"
langchain-community = "^0.3.17"
langgraph = "^0.2.71"
transformers = "^4.48.3"
langchain-openai = "^0.3.5"
fastapi = {extras = ["standard"], version = "^0.115.8"}
uvicorn = "^0.34.0"
pypdf = "^5.3.0"
sentence-transformers = "^3.4.1"
langchain-chroma = "^0.2.1"
gradio = "^5.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
